exercise/L1_word2vec_hpsearch/requirements.txt
nb-black==1.0.7
numpy==1.19.5
torchtext==0.12.0
torchdata==0.3.0
pyyaml==6.0